refactor: move buildHealthCheckOptions.
