+++ server/server_configuration_test.go
@@ -413,7 +413,7 @@ func TestServerMultipleFrontendRules(t *testing.T) {
 	}
 }
 
-func TestServerParseHealthCheckOptions(t *testing.T) {
+func TestServerBuildHealthCheckOptions(t *testing.T) {
 	lb := &testLoadBalancer{}
 	globalInterval := 15 * time.Second
 
@@ -477,81 +477,8 @@ func TestServerParseHealthCheckOptions(t *testing.T) {
 		t.Run(test.desc, func(t *testing.T) {
 			t.Parallel()
 
-			opts := parseHealthCheckOptions(lb, "backend", test.hc, &configuration.HealthCheckConfig{Interval: flaeg.Duration(globalInterval)})
+			opts := buildHealthCheckOptions(lb, "backend", test.hc, &configuration.HealthCheckConfig{Interval: flaeg.Duration(globalInterval)})
 			assert.Equal(t, test.expectedOpts, opts, "health check options")
-		})
-	}
-}
-
-func TestConfigureBackends(t *testing.T) {
-	validMethod := "Drr"
-	defaultMethod := "wrr"
-
-	testCases := []struct {
-		desc               string
-		lb                 *types.LoadBalancer
-		expectedMethod     string
-		expectedStickiness *types.Stickiness
-	}{
-		{
-			desc: "valid load balancer method with sticky enabled",
-			lb: &types.LoadBalancer{
-				Method:     validMethod,
-				Stickiness: &types.Stickiness{},
-			},
-			expectedMethod:     validMethod,
-			expectedStickiness: &types.Stickiness{},
-		},
-		{
-			desc: "valid load balancer method with sticky disabled",
-			lb: &types.LoadBalancer{
-				Method:     validMethod,
-				Stickiness: nil,
-			},
-			expectedMethod: validMethod,
-		},
-		{
-			desc: "invalid load balancer method with sticky enabled",
-			lb: &types.LoadBalancer{
-				Method:     "Invalid",
-				Stickiness: &types.Stickiness{},
-			},
-			expectedMethod:     defaultMethod,
-			expectedStickiness: &types.Stickiness{},
-		},
-		{
-			desc: "invalid load balancer method with sticky disabled",
-			lb: &types.LoadBalancer{
-				Method:     "Invalid",
-				Stickiness: nil,
-			},
-			expectedMethod: defaultMethod,
-		},
-		{
-			desc:           "missing load balancer",
-			lb:             nil,
-			expectedMethod: defaultMethod,
-		},
-	}
-
-	for _, test := range testCases {
-		test := test
-		t.Run(test.desc, func(t *testing.T) {
-			t.Parallel()
-			backend := &types.Backend{
-				LoadBalancer: test.lb,
-			}
-
-			configureBackends(map[string]*types.Backend{
-				"backend": backend,
-			})
-
-			expected := types.LoadBalancer{
-				Method:     test.expectedMethod,
-				Stickiness: test.expectedStickiness,
-			}
-
-			assert.Equal(t, expected, *backend.LoadBalancer)
 		})
 	}
 }
